# Electron/src/swagger/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Spotify Electron Backend API",
    "description": "API created with Python FastAPI to serve          as backend for Spotify Electron music streaming Desktop App",
    "contact": {
      "name": "Antonio Mart\u00ednez Fern\u00e1ndez",
      "url": "https://github.com/AntonioMrtz",
      "email": "antoniomartinezfernandez17@gmail.com"
    },
    "license": {
      "name": "Attribution-NonCommercial-ShareAlike 4.0 International",
      "url": "https://creativecommons.org/licenses/by-nc-sa/4.0/deed.es"
    },
    "version": "2.1.0"
  },
  "paths": {
    "/playlists/{name}": {
      "get": {
        "tags": [
          "Playlists"
        ],
        "summary": "Get Playlist",
        "description": "Get playlist\n\nArgs:\n    name: playlist name\n    token: JWT info",
        "operationId": "get_playlist_playlists__name__get",
        "security": [
          {
            "JWTBearer": []
          }
        ],
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Playlists"
        ],
        "summary": "Update Playlist",
        "description": "Update playlist\n\nArgs:\n    name: playlist name\n    photo: playlist new photo\n    description: playlist new description\n    song_names: playlist new song names. Defaults to Body(...).\n    new_name: playlist new name. Defaults to None.\n    token: JWT info",
        "operationId": "update_playlist_playlists__name__put",
        "security": [
          {
            "JWTBearer": []
          }
        ],
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Name"
            }
          },
          {
            "name": "photo",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Photo"
            }
          },
          {
            "name": "description",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Description"
            }
          },
          {
            "name": "new_name",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "New Name"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "type": "string"
                },
                "title": "Song Names"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Playlists"
        ],
        "summary": "Delete Playlist",
        "description": "Delete playlsit\n\nArgs:\n    name: playlist name",
        "operationId": "delete_playlist_playlists__name__delete",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/playlists/": {
      "post": {
        "tags": [
          "Playlists"
        ],
        "summary": "Create Playlist",
        "description": "Create playlist\n\nArgs:\n    name: playlist name\n    photo: playlist photo\n    description: playlist description\n    song_names: list of song names included in playlist.\n    token: JWT info",
        "operationId": "create_playlist_playlists__post",
        "security": [
          {
            "JWTBearer": []
          }
        ],
        "parameters": [
          {
            "name": "name",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Name"
            }
          },
          {
            "name": "photo",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Photo"
            }
          },
          {
            "name": "description",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Description"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "type": "string"
                },
                "title": "Song Names"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Playlists"
        ],
        "summary": "Get Playlists",
        "description": "Get all playlists",
        "operationId": "get_playlists_playlists__get",
        "security": [
          {
            "JWTBearer": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },
    "/playlists/selected/{names}": {
      "get": {
        "tags": [
          "Playlists"
        ],
        "summary": "Get Selected Playlists",
        "description": "Get selected playlists\n\nArgs:\n    names: playlist names\n    token: JWT info",
        "operationId": "get_selected_playlists_playlists_selected__names__get",
        "security": [
          {
            "JWTBearer": []
          }
        ],
        "parameters": [
          {
            "name": "names",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Names"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/playlists/{name}/songs/": {
      "patch": {
        "tags": [
          "Playlists"
        ],
        "summary": "Add Songs To Playlist",
        "description": "Add songs to playlist\n\nArgs:\n    name: playlist name\n    song_names: song names",
        "operationId": "add_songs_to_playlist_playlists__name__songs__patch",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Name"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "type": "string"
                },
                "title": "Song Names"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Playlists"
        ],
        "summary": "Remove Songs From Playlist",
        "description": "Remove songs from playlist\n\nArgs:\n    name: playlist name\n    song_names: song names",
        "operationId": "remove_songs_from_playlist_playlists__name__songs__delete",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Name"
            }
          },
          {
            "name": "song_names",
            "in": "query",
            "required": true,
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              },
              "title": "Song Names"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/users/{name}/playlists": {
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "Get User Playlists",
        "description": "Get playlists created by the user\n\nArgs:\n    name: user name\n    token: JWT info",
        "operationId": "get_user_playlists_users__name__playlists_get",
        "security": [
          {
            "JWTBearer": []
          }
        ],
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "JWTBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}